Documentation, style, and cleaning
Added documentation, fixed some style concerns. Only keep track of the
last place a bracket was closed instead of a whole stack since that's all
that's needed for the parsing. Rearranged logic for when hanging indent
should be used to accomodate comments on the line before a hanging indent
is used.