Don't throw errors when source code is malformed.
An issue was reported where a new line was entered after a closing
bracket with _no_ matching opening bracket. This causes an error
when popping a non-existing opening bracket off the array of
bracket match locations. This commit simply makes sure there are
elements in the open bracket stack before trying to pop an element
off of it, and provides test cases to illustrate the thrown error
(and prevent regression).