Fix issue when empty string is in the markup (#36)
When an empty string `""` was anywhere in a file, the indentation would
fail because the `checkNextCharForString` variable was being reset on
each line, causing the parsing function to believe indentation was
occurring within a non-terminated string (like a multi-line
triple-quote). This variable instead should be in the context of the
parse function, not just a single line.

The regression was caused by another fix that was made in #34 for
comment characters in a triple-quoted string.

This fixes #35.