Handle missing "decreaseNextIndent" pattern
This pattern was recently removed from language-python.
See https://github.com/atom/language-python/pull/212

This commit handles that scenario, as well as fixes tests surrounding
it. Other tests were adjusted to handle Atom not auto-inserting the
indentation for several scenarios where we are currently using the
`makeNewline` function. The tabs to indent (which are always a multiple
of the editor's tab length) can be passed. This only happens on Atom >
1.23. There is probably a better solution for this in the long term.