Fix logic when cursor is not at the end.
Also simplify code structure by removing the class, and using a plain
function when indenting.

Note: This breaks one test case when using a comment character at the
end of a line where the next indent should be hanging. This is a rare
case, but one I'd like to fix soon.

Fixes #69